pyproject.toml
[project]
name = "gambito-neural-tcc"
version = "0.1.0"
description = "Chess Web APP based in Genai"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.4",
    "fastapi[standard]>=0.121.0",
    "google-genai>=1.49.0",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.1",
    "psycopg2>=2.9.11",
    "pyjwt>=2.10.1",
    "python-chess>=1.999",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "stockfish>=3.28.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
]
